@@ -3,8 +3,8 @@ package com.baeldung.rest.karate;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import com.intuit.karate.junit4.Karate;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.runner.RunWith;
 
 
@@ -18,7 +18,7 @@ public class KarateIntegrationTest {
 
     private static final WireMockServer wireMockServer = new WireMockServer(WireMockConfiguration.options().port(PORT_NUMBER));
 
-    @BeforeClass
+    @BeforeAll
     public static void setUp() {
         wireMockServer.start();
 
@@ -38,7 +38,7 @@ public class KarateIntegrationTest {
 
     }
 
-    @AfterClass
+    @AfterAll
     public static void tearDown() {
         wireMockServer.stop();
     }
