@@ -1,65 +1,69 @@
 package com.baeldung.rest.wiremock.introduction;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
-import static com.github.tomakehurst.wiremock.client.WireMock.containing;
-import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
-import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.matching;
-import static com.github.tomakehurst.wiremock.client.WireMock.post;
-import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlPathEqualTo;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
-import static com.github.tomakehurst.wiremock.client.WireMock.verify;
-import static org.junit.Assert.assertEquals;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.ServerSocket;
-import java.util.Scanner;
-
+import com.github.tomakehurst.wiremock.WireMockServer;
+import com.github.tomakehurst.wiremock.client.WireMock;
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
-import org.junit.Rule;
-import org.junit.Test;
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.ServerSocket;
+import java.util.Scanner;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class JUnitManagedIntegrationTest {
 
     private static final String BAELDUNG_WIREMOCK_PATH = "/baeldung/wiremock";
     private static final String APPLICATION_JSON = "application/json";
     static int port;
-    
+
     static {
-        
+
         try {
             // Get a free port
             ServerSocket s = new ServerSocket(0);
             port = s.getLocalPort();
             s.close();
-            
+
         } catch (IOException e) {
             // No OPS
         }
     }
-    
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(port);
+
+    private WireMockServer wireMockServer;
+
+    @BeforeEach
+    void setup() {
+        wireMockServer = new WireMockServer(WireMockConfiguration.options().port(port));
+        wireMockServer.start();
+        WireMock.configureFor("localhost", port);
+    }
+
+    @AfterEach
+    void teardown() {
+        if (wireMockServer != null) {
+            wireMockServer.stop();
+        }
+    }
 
     @Test
     public void givenJUnitManagedServer_whenMatchingURL_thenCorrect() throws IOException {
-        
+
         stubFor(get(urlPathMatching("/baeldung/.*"))
-                .willReturn(aResponse()
-                        .withStatus(200)
-                        .withHeader("Content-Type", APPLICATION_JSON)
-                        .withBody("\"testing-library\": \"WireMock\"")));
+            .willReturn(aResponse()
+                .withStatus(200)
+                .withHeader("Content-Type", APPLICATION_JSON)
+                .withBody("\"testing-library\": \"WireMock\"")));
 
         CloseableHttpClient httpClient = HttpClients.createDefault();
         HttpGet request = new HttpGet(String.format("http://localhost:%s/baeldung/wiremock", port));
@@ -75,11 +79,11 @@ public class JUnitManagedIntegrationTest {
     @Test
     public void givenJUnitManagedServer_whenMatchingHeaders_thenCorrect() throws IOException {
         stubFor(get(urlPathEqualTo(BAELDUNG_WIREMOCK_PATH))
-                .withHeader("Accept", matching("text/.*"))
-                .willReturn(aResponse()
-                        .withStatus(503)
-                        .withHeader("Content-Type", "text/html")
-                        .withBody("!!! Service Unavailable !!!")));
+            .withHeader("Accept", matching("text/.*"))
+            .willReturn(aResponse()
+                .withStatus(503)
+                .withHeader("Content-Type", "text/html")
+                .withBody("!!! Service Unavailable !!!")));
 
         CloseableHttpClient httpClient = HttpClients.createDefault();
         HttpGet request = new HttpGet(String.format("http://localhost:%s/baeldung/wiremock", port));
@@ -96,11 +100,11 @@ public class JUnitManagedIntegrationTest {
     @Test
     public void givenJUnitManagedServer_whenMatchingBody_thenCorrect() throws IOException {
         stubFor(post(urlEqualTo(BAELDUNG_WIREMOCK_PATH))
-                .withHeader("Content-Type", equalTo(APPLICATION_JSON))
-                .withRequestBody(containing("\"testing-library\": \"WireMock\""))
-                .withRequestBody(containing("\"creator\": \"Tom Akehurst\""))
-                .withRequestBody(containing("\"website\": \"wiremock.org\""))
-                .willReturn(aResponse().withStatus(200)));
+            .withHeader("Content-Type", equalTo(APPLICATION_JSON))
+            .withRequestBody(containing("\"testing-library\": \"WireMock\""))
+            .withRequestBody(containing("\"creator\": \"Tom Akehurst\""))
+            .withRequestBody(containing("\"website\": \"wiremock.org\""))
+            .willReturn(aResponse().withStatus(200)));
 
         InputStream jsonInputStream = this.getClass().getClassLoader().getResourceAsStream("wiremock_intro.json");
         String jsonString = convertInputStreamToString(jsonInputStream);
@@ -113,7 +117,7 @@ public class JUnitManagedIntegrationTest {
         HttpResponse response = httpClient.execute(request);
 
         verify(postRequestedFor(urlEqualTo(BAELDUNG_WIREMOCK_PATH))
-                .withHeader("Content-Type", equalTo(APPLICATION_JSON)));
+            .withHeader("Content-Type", equalTo(APPLICATION_JSON)));
         assertEquals(200, response.getStatusLine().getStatusCode());
     }
 
