@@ -1,46 +1,57 @@
 package com.baeldung.rest.wiremock.scenario;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
-import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
-import static org.junit.Assert.assertEquals;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
+import com.github.tomakehurst.wiremock.WireMockServer;
+import com.github.tomakehurst.wiremock.client.WireMock;
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
+import com.github.tomakehurst.wiremock.stubbing.Scenario;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
-import com.github.tomakehurst.wiremock.stubbing.Scenario;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class WireMockScenarioExampleIntegrationTest {
-    
+
     private static final String THIRD_STATE = "third";
     private static final String SECOND_STATE = "second";
     private static final String TIP_01 = "finally block is not called when System.exit() is called in the try block";
     private static final String TIP_02 = "keep your code clean";
     private static final String TIP_03 = "use composition rather than inheritance";
     private static final String TEXT_PLAIN = "text/plain";
-    
+
     static int port = 9999;
-    
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(port);
-    
+    private WireMockServer wireMockServer;
+
+    @BeforeEach
+    void setup() {
+        wireMockServer = new WireMockServer(WireMockConfiguration.options().port(port));
+        wireMockServer.start();
+        WireMock.configureFor("localhost", port);
+    }
+
+    @AfterEach
+    void teardown() {
+        if (wireMockServer != null) {
+            wireMockServer.stop();
+        }
+    }
+
     @Test
     public void changeStateOnEachCallTest() throws IOException {
         createWireMockStub(Scenario.STARTED, SECOND_STATE, TIP_01);
         createWireMockStub(SECOND_STATE, THIRD_STATE, TIP_02);
         createWireMockStub(THIRD_STATE, Scenario.STARTED, TIP_03);
-        
+
         assertEquals(TIP_01, nextTip());
         assertEquals(TIP_02, nextTip());
         assertEquals(TIP_03, nextTip());
@@ -52,13 +63,13 @@ public class WireMockScenarioExampleIntegrationTest {
             .inScenario("java tips")
             .whenScenarioStateIs(currentState)
             .willReturn(aResponse()
-                    .withStatus(200)
-                    .withHeader("Content-Type", TEXT_PLAIN)
-                    .withBody(responseBody))
+                .withStatus(200)
+                .withHeader("Content-Type", TEXT_PLAIN)
+                .withBody(responseBody))
             .willSetStateTo(nextState)
-            );
+        );
     }
-    
+
     private String nextTip() throws ClientProtocolException, IOException {
         CloseableHttpClient httpClient = HttpClients.createDefault();
         HttpGet request = new HttpGet(String.format("http://localhost:%s/java-tip", port));
@@ -71,5 +82,5 @@ public class WireMockScenarioExampleIntegrationTest {
             return reader.readLine();
         }
     }
-    
+
 }
