@@ -1,8 +1,8 @@
 package com.baeldung.hibernate.entities;
 
-import java.util.List;
-
 import jakarta.persistence.*;
+
+import java.util.List;
 
 @Entity
 public class Department {
