@@ -1,25 +1,6 @@
 package com.baeldung.rest.cucumber;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
-import static com.github.tomakehurst.wiremock.client.WireMock.configureFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.containing;
-import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
-import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.post;
-import static com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
-import static com.github.tomakehurst.wiremock.client.WireMock.verify;
-import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThat;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Scanner;
-
+import com.github.tomakehurst.wiremock.WireMockServer;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import org.apache.http.HttpResponse;
@@ -29,7 +10,15 @@ import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 
-import com.github.tomakehurst.wiremock.WireMockServer;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Scanner;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.junit.Assert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 public class StepDefinition {
@@ -49,9 +38,9 @@ public class StepDefinition {
 
         configureFor("localhost", wireMockServer.port());
         stubFor(post(urlEqualTo(CREATE_PATH))
-                .withHeader("content-type", equalTo(APPLICATION_JSON))
-                .withRequestBody(containing("testing-framework"))
-                .willReturn(aResponse().withStatus(200)));
+            .withHeader("content-type", equalTo(APPLICATION_JSON))
+            .withRequestBody(containing("testing-framework"))
+            .willReturn(aResponse().withStatus(200)));
 
         HttpPost request = new HttpPost("http://localhost:" + wireMockServer.port() + "/create");
         StringEntity entity = new StringEntity(jsonString);
@@ -61,19 +50,19 @@ public class StepDefinition {
 
         assertEquals(200, response.getStatusLine().getStatusCode());
         verify(postRequestedFor(urlEqualTo(CREATE_PATH))
-                .withHeader("content-type", equalTo(APPLICATION_JSON)));
+            .withHeader("content-type", equalTo(APPLICATION_JSON)));
 
         wireMockServer.stop();
     }
 
-//    @When("^users want to get information on the '(.+)' project$")
+    //    @When("^users want to get information on the '(.+)' project$")
     @When("users want to get information on the {string} project")
     public void usersGetInformationOnAProject(String projectName) throws IOException {
         wireMockServer.start();
 
         configureFor("localhost", wireMockServer.port());
         stubFor(get(urlEqualTo("/projects/cucumber")).withHeader("accept", equalTo(APPLICATION_JSON))
-                .willReturn(aResponse().withBody(jsonString)));
+            .willReturn(aResponse().withBody(jsonString)));
 
         HttpGet request = new HttpGet("http://localhost:" + wireMockServer.port() + "/projects/" + projectName.toLowerCase());
         request.addHeader("accept", APPLICATION_JSON);
